feat(authentication): responsiveness for mobile
