+++ src/app/signin/page.tsx
@@ -1,0 +1,11 @@
+import SignInForm from "@/components/SignInForm"
+
+const SignIn = () => {
+  return (
+    <div>
+        <SignInForm />
+    </div>
+  )
+}
+
+export default SignIn
